♻️ refactor: api 도메인 분리
