+++ src/api/caffeine/caffeineInfoApi.ts
@@ -1,11 +1,38 @@
-import apiClient from './apiClient';
+import apiClient from "@/api/apiClient";
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useQueryHooks } from '@/hooks/useQueryHooks';
+import { CaffeineInfo, SleepInfo } from '@/stores/onboardingStore';
+
+const getUserId = () => {
+  const userId = localStorage.getItem('userId');
+  if (!userId) throw new Error('사용자 정보가 없습니다.');
+  return userId;
+};
+
+// ✅ 카페인 설정 등록 (온보딩용)
+export const useSubmitCaffeineInfo = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: async ({ caffeineInfo, sleepInfo }: { caffeineInfo: CaffeineInfo; sleepInfo: SleepInfo }) => {
+      const userId = getUserId();
+      const parsedData = {
+        caffeineSensitivity: caffeineInfo.caffeineSensitivity,
+        averageDailyCaffeineIntake: caffeineInfo.averageDailyCaffeineIntake,
+        frequentDrinkTime: sleepInfo.frequentDrinkTime || "00:00",
+        userFavoriteDrinks: caffeineInfo.userFavoriteDrinks || [],
+      };
+      return apiClient.post(`/api/v1/users/${userId}/caffeine`, parsedData);
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['caffeineInfo'] });
+    },
+  });
+};
 
 // ✅ 사용자 카페인 설정 조회 API
 export const fetchCaffeineInfo = async () => {
-  const userId = localStorage.getItem("userId");
-  if (!userId) throw new Error('사용자 정보가 없습니다.');
+  const userId = getUserId();
   const response = await apiClient.get(`/api/v1/users/${userId}/caffeine`);
   return response.data.data;
 };
@@ -58,7 +85,7 @@ export const useUpdateCaffeineInfo = () => {
   
     const mutation = useMutation({
       mutationFn: async (newData: Record<string, any>) => {
-        const prev = queryClient.getQueryData<Record<string, any>>(['caffeineSettings']);
+        const prev = queryClient.getQueryData<Record<string, any>>(['caffeineInfo']);
         if (!prev) {
           return updateCaffeineInfo(newData);
         }
@@ -69,7 +96,7 @@ export const useUpdateCaffeineInfo = () => {
         return updateCaffeineInfo(diffData);
       },
       onSuccess: () => {
-        queryClient.invalidateQueries({ queryKey:['caffeineSettings']});
+        queryClient.invalidateQueries({ queryKey:['caffeineInfo']});
       },
       onError: (error: any) => {
         console.error('카페인 설정 수정 중 오류:', error);
