+++ src/components/report/DropdownSelector.tsx
@@ -30,7 +30,7 @@ const DropdownSelector: React.FC<DropdownSelectorProps> = ({
           className="
             flex items-center justify-between
             w-28 px-4 py-2
-            bg-white border border-gray-200 rounded-lg
+            bg-white border border-[#dadcdf] rounded-lg
             text-[#333333] font-medium
             hover:bg-gray-50 focus:outline-none
           "
