+++ src/pages/HomePage.tsx
@@ -9,8 +9,8 @@ import BannerImage2 from '@/assets/Banner02.png';
 import BannerImage3 from '@/assets/Banner03.png';
 import CaffeineBottomSheet from '@/components/caffeine/CaffeineBottomSheet';
 import type { CaffeineRecordInput } from '@/components/caffeine/CaffeineDetailForm';
-import { recordCaffeineIntake } from '@/api/caffeineApi';
-import { useDailyCaffeineReport } from '@/api/dailyReportApi';
+import { recordCaffeineIntake } from '@/api/caffeine/caffeineApi';
+import { useDailyCaffeineReport } from '@/api/home/dailyReportApi';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
 import AlertModal from '@/components/common/AlertModal';
 import { Info, AlertTriangle } from 'lucide-react';
