+++ src/pages/diary/DiaryPage.tsx
@@ -6,12 +6,12 @@ import CaffeineList from '@/components/diary/CaffeineList';
 import { useNavigate } from 'react-router-dom';
 import CaffeineBottomSheet from '@/components/caffeine/CaffeineBottomSheet';
 import type { CaffeineRecordInput } from '@/components/caffeine/CaffeineDetailForm';
-import { useCalendar } from '@/api/calendarApi';
+import { useCalendar } from '@/api/diary/calendarApi';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
 import EmptyState from '@/components/common/EmptyState';
 import { AlertTriangle } from "lucide-react";
-import { useDailyIntake } from '@/api/calendarListApi';
-import { recordCaffeineIntake } from '@/api/caffeineApi';
+import { useDailyIntake } from '@/api/diary/calendarListApi';
+import { recordCaffeineIntake } from '@/api/caffeine/caffeineApi';
 import AlertModal from '@/components/common/AlertModal';
 import SectionCard from '@/components/common/SectionCard';
 import { formatDate } from '@/utils/date';
