+++ src/pages/mypage/MypageEditPage.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import Header from '@/components/common/Header';
 import HealthInfoEditor from '@/components/mypage/HealthInfoEditor';
 import { useNavigate } from 'react-router-dom';
-import { useUpdateHealthInfo } from '@/api/healthInfoApi';
-import { useUpdateCaffeineInfo } from '@/api/caffeineInfoApi';
+import { useUpdateHealthInfo } from '@/api//health/healthInfoApi';
+import { useUpdateCaffeineInfo } from '@/api/caffeine/caffeineInfoApi';
 import AlertModal from '@/components/common/AlertModal';
 import { Info } from 'lucide-react';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
