+++ src/pages/mypage/MypagePage.tsx
@@ -5,9 +5,9 @@ import { useNavigate } from 'react-router-dom';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
 import AlertModal from '@/components/common/AlertModal';
 import { AlertCircle, AlertTriangle } from 'lucide-react';
-import { useUserProfile } from '@/api/profileApi';
-import { useLogout } from '@/api/LogoutApi';
-import { useDeleteUser } from '@/api/deletUserApi';
+import { useUserProfile } from '@/api/mypage/profileApi';
+import { useLogout } from '@/api/mypage/LogoutApi';
+import { useDeleteUser } from '@/api/mypage/deletUserApi';
 import { useState } from 'react';
 import EmptyState from '@/components/common/EmptyState';
 import SectionCard from '@/components/common/SectionCard';
