+++ src/pages/report/ReportPage.tsx
@@ -9,10 +9,10 @@ import ReportMessage from '@/components/report/ReportMessage';
 import { useNavigate } from 'react-router-dom';
 import CaffeineBottomSheet from '@/components/caffeine/CaffeineBottomSheet';
 import type { CaffeineRecordInput } from '@/components/caffeine/CaffeineDetailForm';
-import { useWeeklyReport } from '@/api/weeklyReportApi';
-import { useMonthlyReport } from '@/api/monthlyReportApi';
-import { useYearlyReport } from '@/api/yearlyReportApi';
-import { recordCaffeineIntake } from '@/api/caffeineApi';
+import { useWeeklyReport } from '@/api/report/weeklyReportApi';
+import { useMonthlyReport } from '@/api/report/monthlyReportApi';
+import { useYearlyReport } from '@/api/report/yearlyReportApi';
+import { recordCaffeineIntake } from '@/api/caffeine/caffeineApi';
 import LoadingSpinner from '@/components/common/LoadingSpinner';
 import EmptyState from '@/components/common/EmptyState';
 import AlertModal from '@/components/common/AlertModal';
